Lots of timer stuff.
Fixed a "feature" in pc_system.cc with setting timers to small values
that can cause bochs to hang.

Significantly improved the performance of the new PIT.
It's probably ready to become the default now.

Added a preliminary implementation of the slowdown timer
that Bryce and I had talked about.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@1169 5c200c8f-f062-490a-aca5-964d1c9e8edb