Fix problem with PIT when the machine is being really slow.
This means that the REALTIME_PIT now only works on machines
that are capable of sustaining 150000 IPS, in aggregate.  This
seems a reasonable requirement, as even my 200Mhz PPC with 16MB
of memory can sustain that.

This is now probably ready for primetime, so I need to get it
added to the configure options.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@2447 5c200c8f-f062-490a-aca5-964d1c9e8edb