There's a bug in the repeated IO & mem copy speedups.  I added --enable-repeat-speedups with default to disabled. Reconfigure/recompile and the speedup code will be #ifdef'd out for now.  It manifested as junk written to the VGA screen while booting/running Windows.
Also made some more mods to the main cpu loop.  Moved the
handling of EXT/errorno outside the main loop, much like
the extra EIP/ESP commits were moved, for a little better
performance.

I changed the fetch_ptr/bytesleft method of fetching to
a slightly different model, which calculates a window
for which EIP will be valid (land on the current page),
and a bias which when applied to EIP will be from
0..upper_page_limit.  Speed is about the same for either
method, but a pseudo-op/threaded-interpreter will plug
in better with this and be faster.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@2755 5c200c8f-f062-490a-aca5-964d1c9e8edb