I rehashed the way the paging code stores protection bits, so that a compare of the current access could be done more efficiently against the cached values, both in the normal paging routines, and in the accelerated code in access.cc.
This cut down the amount of code path needed to get to
direct use of a host address nicely, and speed definitely
got a boost as a result, especially if you use the
--enable-guest2host-tlb option.

The CR0.WP flag was a real pain, because it imparts
a complication on the way protections work.  Fortunately
it's not a high-change flag, so I just base the new
cached info on the current CR0.WP value, and dump
the TLB cache when it changes.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@2792 5c200c8f-f062-490a-aca5-964d1c9e8edb