I extended the guest-to-host TLB acceleration across the Read-Modify-Write instructions.  The first read phase stores the host pointer in the "pages" field if a direct use pointer is available.  The Write phase first checks if a pointer was issued and uses it for a direct write if available.
I chose the "pages" field since it needs to be checked by the
write_RMW_virtual variants anyways and thus needs to be
cached anyways.

Mostly the mods where to access.cc, but I did also macro-ize
the calls to write_RMW_virtual...() in files which use it
and cpu.h.  Right now, the macro is just a straight pass-through.
I tried expanding it to a quick initial check for the pointer
availability to do the write in-place, with a function call
as a fall-back.  That didn't seemed to matter at all.

Booting is not helped by this really.  The upper bound of
the gain is 5 or 6%, and that's only if you have a loop that
looks like:

label:
  add [eax], ebx   ;; mega read-modify-write instruction
  jmp label        ;; intensive loop.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@2848 5c200c8f-f062-490a-aca5-964d1c9e8edb