Updated accessing of modrm/sib addressing information to use accessors.  This lets me work on compressing the size of fetch-decode structure (now called bxInstruction_c).
I've reduced it down to about 76 bytes.  We should be able
to do much better soon.  I needed the abstraction of the
accessors, so I have a lot of freedom to re-arrange things
without making massive future changes.

Lost a few percent of performance in these mods, but my
main focus was to get the abstraction.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@3011 5c200c8f-f062-490a-aca5-964d1c9e8edb