Implemented decoding tables for decoding of MMX/SSE/SSE instruction dependent to prefixes (none, 0x66, 0xf2 or 0xf3)
Introduced a new way to solve the problem with "instructions not supported
in current configuration". If CPU decodes instruction that not unsupported in current
configuration (for example an SSE instruction when cpu-level=5) it must report
UD exception before(without) fetching of extra opcode bytes such as modrm or imm.

Wait, In this step the code still fails compilation ;)


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/branches/unlabeled-1.27.4/bochs@3397 5c200c8f-f062-490a-aca5-964d1c9e8edb