Latest round of patches/fixups to get 64 bit emulation further.
This is an interim update to allow others to test.

We have userland code running!!! (up to a point)

Able to start executing "sash" as /sbin/init in userland from linux 64 bit
kernel until it crashes trying to access a null pointer.  No kernel panics
though, just a segfault loop.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@3413 5c200c8f-f062-490a-aca5-964d1c9e8edb