Temporary tweak to reinstate a change that disappeared when sse2.cc was removed.
The 64 bit variant of MOVNTI was not decoded.  The proper fix for this is to work on
fetchdecode64.cc to call a 64 bit variant of SSE instructions or fail it with a
invalid op.  A careful check needs to be done with the AMD manuals to determine if
there are any other SSE instructions that have a special 64 bit decoding.


git-svn-id: https://bochs.svn.sourceforge.net/svnroot/bochs/trunk/bochs@4543 5c200c8f-f062-490a-aca5-964d1c9e8edb